requirements.txt
absl-py==2.1.0
actionlib==1.14.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
angles==1.9.13
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
autobahn==17.10.1
Automat==0.8.0
babel==2.16.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
cachetools==5.5.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2024.8.30
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
comm==0.2.2
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
docutils==0.16
dulwich==0.19.15
dynamic-reconfigure==1.7.3
easydict==1.13
EasyPySpin==2.0.1
einops==0.8.0
empy==3.3.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastimport==0.9.8
fastjsonschema==2.20.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gdown==5.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gpg==1.13.1
grpcio==1.66.1
h5py==3.11.0
httplib2==0.14.0
huggingface-hub==0.25.1
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
imageio==2.35.1
importlib-metadata==8.4.0
importlib-resources==6.4.4
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-packaging==0.12.3
jupyter-server==2.14.2
jupyter-server-fileid==0.9.2
jupyter-server-terminals==0.5.3
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.8
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lap==0.4.0
lapx==0.5.10.post1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.4
loguru==0.7.2
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
message-filters==1.16.0
mistune==3.0.2
motmetrics==1.4.0
mpi4py==3.0.3
mpmath==1.3.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
ninja==1.11.1.1
nose==1.3.7
notebook==6.5.7
notebook-shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
onnxruntime==1.8.0
open3d==0.13.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
prettytable==3.11.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.22
pycryptodomex==3.6.1
pycups==1.9.73
pydensecrf==1.0
pydot==1.4.1
PyGithub==1.43.7
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-gitlab==2.0.1
python-gnupg==0.4.5
python-json-logger==2.0.7
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==2.0.0
rosbag==1.17.0
rosbags==0.9.23
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==0.9.1
rosdistro-modules==0.9.1
rosgraph==1.16.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospkg-modules==1.5.1
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.20.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rviz==1.14.25
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.3
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
soupsieve==2.6
spinnaker-python==3.1.0.79
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.13.2
topic-tools==1.16.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.12.2
tzdata==2024.1
u-msgpack-python==2.1
ultralytics==8.1.34
urllib3==2.2.2
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.18
xmltodict==0.13.0
y-py==0.6.2
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.20.2
zope.interface==4.7.1
zstandard==0.23.0
